feat: add configuration files and update ADS1248 driver for temperature conversion

+++ RTD_firmware/Core/Inc/ADS1248.h
@@ -20,8 +20,10 @@
 void ADS124X_init(SPI_HandleTypeDef *hspi,
         GPIO_TypeDef *GPIO_START, uint16_t PIN_START,
         GPIO_TypeDef *GPIO_CS, uint16_t PIN_CS);
-int32_t Temperature(SPI_HandleTypeDef *hspi,
+
+float Temperature(SPI_HandleTypeDef *hspi,
         GPIO_TypeDef *GPIO_CS, uint16_t PIN_CS);
+
 float RTD_Converter(int32_t result);
 
 #endif /* INC_ADS1248_H_ */
